Mejorando la estructura
