+++ src/Routes/nota.md
@@ -1,0 +1,2 @@
+# Route o Navigation
+ Configuración de la navegación de la aplicación, utilizando bibliotecas como React Navigation.
