+++ src/auth/key/security
@@ -1,0 +1,5 @@
+export default {
+    // encryptionKey: 'colegio9p2ik7h&8^d#)$1hw7(t1llxnaqwe2%rzd=icueji%&mpbwks3app'
+    // encryptionKey: '0a703b08d995357c4bd48173c9dcdc25'
+  };
+
