+++ src/screens/nota.md
@@ -1,0 +1,2 @@
+# screens: 
+Las pantallas de la aplicación, cada archivo aquí generalmente corresponde a una pantalla de la aplicación.
