+++ src/services/url.jsx
@@ -1,0 +1,1 @@
+export const BASE_URL= 'https://7b79-181-208-46-146.ngrok-free.app';
